+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Issue #624: Text on line breaks is now correctly highlighted.
 * Issue #708: Removing large portions of text no longer freezes the app.
+* Issue #732: User-defined width of sidebar is now restored on load.
 
 ## [2.1.1] - 2021-12-24